LayoutTests/inspector/debugger/tail-deleted-frames-from-vm-entry-expected.txt

Testing that we keep around tail deleted frames that are entry frames.

Starting Test


------------------------------------
Hit breakpoint at line: 7, column: 4
------------------------------------
Expected frame: {"functionName":"bar","scope":["i",0],"isTailDeleted":false}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: false
Expected frame: {"functionName":"bar","scope":["i",1],"isTailDeleted":true}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: true
Expected frame: {"functionName":"bar","scope":["i",2],"isTailDeleted":true}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: true
Expected frame: {"functionName":"bar","scope":["i",3],"isTailDeleted":true}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: true
Expected frame: {"functionName":"bar","scope":["i",4],"isTailDeleted":true}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: true
Expected frame: {"functionName":"bar","scope":["i",5],"isTailDeleted":true}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: true
Expected frame: {"functionName":"bar","scope":["i",6],"isTailDeleted":true}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: true
Expected frame: {"functionName":"bar","scope":["i",7],"isTailDeleted":true}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: true
Expected frame: {"functionName":"bar","scope":["i",8],"isTailDeleted":true}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: true
Expected frame: {"functionName":"bar","scope":["i",9],"isTailDeleted":true}
PASS: Function name: bar is correct.
PASS: Tail deleted expectation correct: true
Expected frame: {"functionName":"timeout","scope":["foo",25],"isTailDeleted":true}
PASS: Function name: timeout is correct.
PASS: Tail deleted expectation correct: true
Looking at frame number: 0
    variable 'i': {"_type":"number","_description":"0","_hasChildren":false,"_value":0}
PASS: Variable is a number.
PASS: Found scope value: 0
PASS: Did not find variable we were looking for: i
Looking at frame number: 1
    variable 'i': {"_type":"number","_description":"1","_hasChildren":false,"_value":1}
PASS: Variable is a number.
PASS: Found scope value: 1
PASS: Did not find variable we were looking for: i
Looking at frame number: 2
    variable 'i': {"_type":"number","_description":"2","_hasChildren":false,"_value":2}
PASS: Variable is a number.
PASS: Found scope value: 2
PASS: Did not find variable we were looking for: i
Looking at frame number: 3
    variable 'i': {"_type":"number","_description":"3","_hasChildren":false,"_value":3}
PASS: Variable is a number.
PASS: Found scope value: 3
PASS: Did not find variable we were looking for: i
Looking at frame number: 4
    variable 'i': {"_type":"number","_description":"4","_hasChildren":false,"_value":4}
PASS: Variable is a number.
PASS: Found scope value: 4
PASS: Did not find variable we were looking for: i
Looking at frame number: 5
    variable 'i': {"_type":"number","_description":"5","_hasChildren":false,"_value":5}
PASS: Variable is a number.
PASS: Found scope value: 5
PASS: Did not find variable we were looking for: i
Looking at frame number: 6
    variable 'i': {"_type":"number","_description":"6","_hasChildren":false,"_value":6}
PASS: Variable is a number.
PASS: Found scope value: 6
PASS: Did not find variable we were looking for: i
Looking at frame number: 7
    variable 'i': {"_type":"number","_description":"7","_hasChildren":false,"_value":7}
PASS: Variable is a number.
PASS: Found scope value: 7
PASS: Did not find variable we were looking for: i
Looking at frame number: 8
    variable 'i': {"_type":"number","_description":"8","_hasChildren":false,"_value":8}
PASS: Variable is a number.
PASS: Found scope value: 8
PASS: Did not find variable we were looking for: i
Looking at frame number: 9
    variable 'i': {"_type":"number","_description":"9","_hasChildren":false,"_value":9}
PASS: Variable is a number.
PASS: Found scope value: 9
PASS: Did not find variable we were looking for: i
Looking at frame number: 10
    variable 'foo': {"_type":"number","_description":"25","_hasChildren":false,"_value":25}
PASS: Variable is a number.
PASS: Found scope value: 25
PASS: Did not find variable we were looking for: foo
Tests done
